Make float.round return Float, not Int
Type conversions belong exclusively in to_int/to_float. The float
module's round/ceil/floor are rounding operations that stay in the
Float domain. Use float.to_int(float.round(x)) for integer conversion.

+++ src/typechecker.rs
@@ -1761,10 +1761,10 @@ impl TypeChecker {
             Box::new(Type::Generic("Result".into(), vec![Type::Float, Type::String])),
         )));
 
-        // float.round: (Float) -> Int
+        // float.round: (Float) -> Float
         env.define("float.round".into(), Scheme::mono(Type::Fun(
             vec![Type::Float],
-            Box::new(Type::Int),
+            Box::new(Type::Float),
         )));
 
         // float.ceil: (Float) -> Float
